requirements.txt
django>=3.0.7
django-cors-headers==3.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-allauth==0.41.0
django-filter==2.2.0
dj-rest-auth==1.0.9
python-decouple==3.3
dj-database-url==0.5.0
whitenoise==5.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
